Refactors service catalog and discovery client code and moves them into the kclient package.
It also refactors the operator code.
