+++ pkg/odo/cli/component/unlink.go
@@ -60,7 +60,7 @@ func (o *UnlinkOptions) Complete(name string, cmd *cobra.Command, args []string)
 		return err
 	}
 
-	o.csvSupport, err = o.Client.IsCSVSupported()
+	o.csvSupport, err = o.Client.GetKubeClient().IsCSVSupported()
 	if err != nil {
 		return err
 	}
